Исправление ДЗ
